@@ -1,60 +1,43 @@
 {
-    "name": "angular-material-example",
-    "version": "0.0.1",
-    "license": "Apache-2.0",
+    "name": "ng15",
+    "version": "0.0.0",
     "scripts": {
         "ng": "ng",
         "start": "ng serve",
         "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "postinstall": "ngcc"
+        "watch": "ng build --watch --configuration development",
+        "test": "ng test"
     },
-    "author": "Philip Washington Sorst <philip@sorst.net> (https://sorst.net)",
     "private": true,
     "dependencies": {
-        "@angular/animations": "^10.2.1",
-        "@angular/cdk": "^10.2.6",
-        "@angular/common": "^10.2.1",
-        "@angular/compiler": "^10.2.1",
-        "@angular/core": "^10.2.1",
-        "@angular/forms": "^10.2.1",
-        "@angular/material": "^10.2.6",
-        "@angular/platform-browser": "^10.2.1",
-        "@angular/platform-browser-dynamic": "^10.2.1",
-        "@angular/pwa": "0.1002.0",
-        "@angular/router": "^10.2.1",
-        "@angular/service-worker": "^10.2.1",
-        "@dontdrinkandroot/angular-extensions": "^0.0",
-        "@dontdrinkandroot/angular-material-extensions": "^0.0",
-        "bootstrap": "^4.5.3",
-        "core-js": "^2.5.3",
-        "faker": "^5.1.0",
-        "material-design-icons": "^3.0.1",
-        "rxjs": "^6.6.3",
-        "tslib": "^2.0.0",
-        "zone.js": "^0.10.3"
+        "@angular/animations": "^15.2.0",
+        "@angular/cdk": "15.2.3",
+        "@angular/common": "^15.2.0",
+        "@angular/compiler": "^15.2.0",
+        "@angular/core": "^15.2.0",
+        "@angular/forms": "^15.2.0",
+        "@angular/material": "15.2.3",
+        "@angular/platform-browser": "^15.2.0",
+        "@angular/platform-browser-dynamic": "^15.2.0",
+        "@angular/router": "^15.2.0",
+        "@dontdrinkandroot/ngx-extensions": "^0.3.1",
+        "@dontdrinkandroot/ngx-material-extensions": "^0.3.1",
+        "@faker-js/faker": "^7.6.0",
+        "rxjs": "~7.8.0",
+        "tslib": "^2.3.0",
+        "zone.js": "~0.12.0"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.1002.0",
-        "@angular/cli": "^10.2.0",
-        "@angular/compiler-cli": "^10.2.1",
-        "@angular/language-service": "^10.2.1",
-        "@types/jasmine": "~3.5.0",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "^12.11.1",
-        "codelyzer": "^6.0.0",
-        "jasmine-core": "~3.6.0",
-        "jasmine-spec-reporter": "~5.0.0",
-        "karma": "~5.0.0",
+        "@angular-devkit/build-angular": "^15.2.4",
+        "@angular/cli": "~15.2.4",
+        "@angular/compiler-cli": "^15.2.0",
+        "@types/jasmine": "~4.3.0",
+        "jasmine-core": "~4.5.0",
+        "karma": "~6.4.0",
         "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage-istanbul-reporter": "~3.0.2",
-        "karma-jasmine": "~4.0.0",
-        "karma-jasmine-html-reporter": "^1.5.0",
-        "protractor": "~7.0.0",
-        "ts-node": "~8.3.0",
-        "tslint": "~6.1.0",
-        "typescript": "~4.0.5"
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.0.0",
+        "typescript": "~4.9.4"
     }
 }
