@@ -1,80 +1,63 @@
+import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
-import {APP_INITIALIZER, NgModule} from '@angular/core';
 
 import {AppRoutingModule} from './app-routing.module';
-
 import {AppComponent} from './app.component';
-import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {MatAutocompleteModule} from "@angular/material/autocomplete";
-import {MatButtonModule} from "@angular/material/button";
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {MatListModule} from "@angular/material/list";
+import {DdrMaterialExtensionsModule} from '@dontdrinkandroot/ngx-material-extensions';
+import {DdrExtensionsModule} from '@dontdrinkandroot/ngx-extensions';
 import {MatCardModule} from "@angular/material/card";
-import {MatChipsModule} from "@angular/material/chips";
+import {ChipComponent} from "./components/chip/chip.component";
+import {DragAndDropDetailComponent} from "./examples/drag-and-drop-list/drag-and-drop-detail.component";
+import {DragAndDropListComponent} from "./examples/drag-and-drop-list/drag-and-drop-list.component";
+import {NotFoundComponent} from "./common/not-found.component";
+import {TypographyComponent} from "./typography/typography.component";
+import {TitleComponent} from "./title/title.component";
+import {MatToolbarModule} from "@angular/material/toolbar";
 import {MatFormFieldModule} from "@angular/material/form-field";
 import {MatIconModule} from "@angular/material/icon";
-import {MatListModule} from "@angular/material/list";
 import {MatMenuModule} from "@angular/material/menu";
-import {MatSidenavModule} from "@angular/material/sidenav";
-import {MatToolbarModule} from "@angular/material/toolbar";
-import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {DragDropModule} from '@angular/cdk/drag-drop';
-import {NotFoundComponent} from './common/not-found.component';
-import {ServiceWorkerModule} from '@angular/service-worker';
-import {environment} from '../environments/environment';
-import {DragAndDropListComponent} from './examples/drag-and-drop-list/drag-and-drop-list.component';
-import {DragAndDropDetailComponent} from './examples/drag-and-drop-list/drag-and-drop-detail.component';
-import {ChipComponent} from './components/chip/chip.component';
-import {InitService} from './init/init.service';
-import {TypographyComponent} from './typography/typography.component';
-import {TitleComponent} from './title/title.component';
-import {ScrollingModule} from '@angular/cdk/scrolling';
-import {DdrAngularMaterialExtensionsModule} from '@dontdrinkandroot/angular-material-extensions';
-
-export function initServiceFactory(initService: InitService): Function
-{
-    return () => initService.initialize();
-}
+import {DragDropModule} from "@angular/cdk/drag-drop";
+import {ScrollingModule} from "@angular/cdk/scrolling";
+import {MatChipsModule} from "@angular/material/chips";
+import {ReactiveFormsModule} from "@angular/forms";
+import {MatOptionModule} from "@angular/material/core";
+import {MatAutocompleteModule} from "@angular/material/autocomplete";
+import {MatButtonModule} from "@angular/material/button";
 
 @NgModule({
     declarations: [
         AppComponent,
-        TypographyComponent,
-        TitleComponent,
         ChipComponent,
-        NotFoundComponent,
+        DragAndDropDetailComponent,
         DragAndDropListComponent,
-        DragAndDropDetailComponent
+        NotFoundComponent,
+        TitleComponent,
+        TypographyComponent
     ],
     imports: [
         BrowserModule,
-        BrowserAnimationsModule,
-        FormsModule,
-        ReactiveFormsModule,
         AppRoutingModule,
-        MatSidenavModule,
-        MatIconModule,
-        MatListModule,
-        MatToolbarModule,
+        BrowserAnimationsModule,
+        DdrExtensionsModule,
+        DdrMaterialExtensionsModule,
+        DragDropModule,
+        MatAutocompleteModule,
         MatButtonModule,
         MatCardModule,
-        MatAutocompleteModule,
         MatChipsModule,
         MatFormFieldModule,
+        MatIconModule,
+        MatListModule,
         MatMenuModule,
-        ScrollingModule,
-        DragDropModule,
-        DdrAngularMaterialExtensionsModule,
-        ServiceWorkerModule.register('ngsw-worker.js', {enabled: environment.production})
+        MatOptionModule,
+        MatToolbarModule,
+        ReactiveFormsModule,
+        ScrollingModule
     ],
-    providers: [
-        {
-            provide: APP_INITIALIZER,
-            useFactory: initServiceFactory,
-            deps: [InitService],
-            multi: true
-        },
-    ],
+    providers: [],
     bootstrap: [AppComponent]
 })
-export class AppModule
-{
+export class AppModule {
 }
