@@ -1,6 +1,7 @@
 import {Pagination} from './pagination';
 
-export class PaginatedResult<T> extends Array<T>
-{
-    public pagination: Pagination;
+export class PaginatedResult<T> extends Array<T> {
+    constructor(public readonly pagination: Pagination) {
+        super();
+    }
 }
