@@ -1,28 +1,27 @@
 import {Injectable} from '@angular/core';
 import {User} from './user';
-import * as faker from 'faker';
+import {faker} from '@faker-js/faker';
 
 @Injectable({
     providedIn: 'root'
 })
 export class UserGeneratorService
 {
-    public generate(withUuid: boolean = false): User
-    {
-        let user = new User();
-        if (withUuid) {
-            user.id = this.generateUuid();
+    public generate(): User {
+        let firstName = faker.name.firstName();
+        let lastName = faker.name.lastName();
+        return {
+            id: this.generateUuid(),
+            firstName: firstName,
+            lastName: lastName,
+            userName: lastName.toLocaleLowerCase() + '.' + firstName.toLocaleLowerCase(),
+            avatarUrl: faker.image.avatar(),
+            sortKey: 0
         }
-        user.firstName = faker.name.firstName();
-        user.lastName = faker.name.lastName();
-        user.userName = user.lastName.toLocaleLowerCase() + '.' + user.firstName.toLocaleLowerCase();
-        user.avatarUrl = faker.image.avatar();
-
-        return user;
     }
 
     public generateUuid(): string
     {
-        return faker.random.uuid();
+        return faker.datatype.uuid();
     }
 }
