@@ -1,9 +1,8 @@
-export class User
-{
-    public id: string;
-    public firstName: string;
-    public lastName: string;
-    public userName: string;
-    public avatarUrl: string;
-    public sortKey: number;
+export interface User {
+    id: string;
+    firstName: string;
+    lastName: string;
+    userName: string;
+    avatarUrl: string;
+    sortKey: number;
 }
